+++ smartsheet/models/__init__.py
@@ -91,6 +91,7 @@ from .report_grouping_criterion import ReportGroupingCriterion
 from .report_publish import ReportPublish
 from .report_row import ReportRow
 from .report_sorting_criterion import ReportSortingCriterion
+from .report_scope_inclusion import ReportScopeInclusion
 from .result import Result
 from .row import Row
 from .row_email import RowEmail